+++ tests/modbus_client_channel_test.cc
@@ -8,6 +8,7 @@ TEST(modbus_client_channel, read_bits) {
   tk_object_t* obj = conf_json_load("file://./tests/testdata/read_bits.json", FALSE);
   conf_doc_t* conf = conf_obj_get_doc(obj);
 
+  return_if_fail(conf != NULL);
   modbus_client_channel_t* channel = modbus_client_channel_create(conf->root);
   ASSERT_EQ(channel->unit_id, 1);
   ASSERT_STREQ(channel->name, "read_bits");
@@ -21,7 +22,8 @@ TEST(modbus_client_channel, read_bits) {
   ASSERT_EQ(channel->read_buffer != NULL, true);
   ASSERT_EQ(channel->write_buffer == NULL, true);
   ASSERT_EQ(channel->keep_last_value_if_read_failed, TRUE);
-
+  
+  modbus_client_channel_destroy(channel);
   conf_doc_destroy(conf);
 }
 
@@ -29,6 +31,7 @@ TEST(modbus_client_channel, read_input_bits) {
   tk_object_t* obj = conf_json_load("file://./tests/testdata/read_input_bits.json", FALSE);
   conf_doc_t* conf = conf_obj_get_doc(obj);
 
+  return_if_fail(conf != NULL);
   modbus_client_channel_t* channel = modbus_client_channel_create(conf->root);
   ASSERT_EQ(channel->unit_id, 1);
   ASSERT_STREQ(channel->name, "read_input_bits");
@@ -43,6 +46,7 @@ TEST(modbus_client_channel, read_input_bits) {
   ASSERT_EQ(channel->write_buffer == NULL, true);
   ASSERT_EQ(channel->keep_last_value_if_read_failed, TRUE);
 
+  modbus_client_channel_destroy(channel);
   conf_doc_destroy(conf);
 }
 
@@ -50,6 +54,7 @@ TEST(modbus_client_channel, read_registers) {
   tk_object_t* obj = conf_json_load("file://./tests/testdata/read_registers.json", FALSE);
   conf_doc_t* conf = conf_obj_get_doc(obj);
 
+  return_if_fail(conf != NULL);
   modbus_client_channel_t* channel = modbus_client_channel_create(conf->root);
   ASSERT_EQ(channel->unit_id, 2);
   ASSERT_STREQ(channel->name, "read_registers");
@@ -64,6 +69,7 @@ TEST(modbus_client_channel, read_registers) {
   ASSERT_EQ(channel->write_buffer == NULL, true);
   ASSERT_EQ(channel->keep_last_value_if_read_failed, FALSE);
 
+  modbus_client_channel_destroy(channel);
   conf_doc_destroy(conf);
 }
 
@@ -71,6 +77,7 @@ TEST(modbus_client_channel, read_input_registers) {
   tk_object_t* obj = conf_json_load("file://./tests/testdata/read_input_registers.json", FALSE);
   conf_doc_t* conf = conf_obj_get_doc(obj);
 
+  return_if_fail(conf != NULL);
   modbus_client_channel_t* channel = modbus_client_channel_create(conf->root);
   ASSERT_EQ(channel->unit_id, 2);
   ASSERT_STREQ(channel->name, "read_input_registers");
@@ -85,6 +92,7 @@ TEST(modbus_client_channel, read_input_registers) {
   ASSERT_EQ(channel->write_buffer == NULL, true);
   ASSERT_EQ(channel->keep_last_value_if_read_failed, FALSE);
 
+  modbus_client_channel_destroy(channel);
   conf_doc_destroy(conf);
 }
 
@@ -93,6 +101,7 @@ TEST(modbus_client_channel, write_1_bit) {
   tk_object_t* obj = conf_json_load("file://./tests/testdata/write_1_bit.json", FALSE);
   conf_doc_t* conf = conf_obj_get_doc(obj);
 
+  return_if_fail(conf != NULL);
   modbus_client_channel_t* channel = modbus_client_channel_create(conf->root);
   ASSERT_EQ(channel->unit_id, 1);
   ASSERT_STREQ(channel->name, "write_1_bit");
@@ -106,6 +115,7 @@ TEST(modbus_client_channel, write_1_bit) {
   ASSERT_EQ(channel->write_buffer != NULL, true);
   ASSERT_EQ(channel->read_buffer == NULL, true);
 
+  modbus_client_channel_destroy(channel);
   conf_doc_destroy(conf);
 }
 
@@ -113,6 +123,7 @@ TEST(modbus_client_channel, write_bits) {
   tk_object_t* obj = conf_json_load("file://./tests/testdata/write_bits.json", FALSE);
   conf_doc_t* conf = conf_obj_get_doc(obj);
 
+  return_if_fail(conf != NULL);
   modbus_client_channel_t* channel = modbus_client_channel_create(conf->root);
   ASSERT_EQ(channel->unit_id, 1);
   ASSERT_STREQ(channel->name, "write_bits");
@@ -126,6 +137,7 @@ TEST(modbus_client_channel, write_bits) {
   ASSERT_EQ(channel->write_buffer != NULL, true);
   ASSERT_EQ(channel->read_buffer == NULL, true);
 
+  modbus_client_channel_destroy(channel);
   conf_doc_destroy(conf);
 }
 
@@ -134,6 +146,7 @@ TEST(modbus_client_channel, write_1_register) {
   tk_object_t* obj = conf_json_load("file://./tests/testdata/write_1_register.json", FALSE);
   conf_doc_t* conf = conf_obj_get_doc(obj);
 
+  return_if_fail(conf != NULL);
   modbus_client_channel_t* channel = modbus_client_channel_create(conf->root);
   ASSERT_EQ(channel->unit_id, 1);
   ASSERT_STREQ(channel->name, "write_1_register");
@@ -154,7 +167,7 @@ TEST(modbus_client_channel, write_1_register) {
 TEST(modbus_client_channel, write_registers) {
   tk_object_t* obj = conf_json_load("file://./tests/testdata/write_registers.json", FALSE);
   conf_doc_t* conf = conf_obj_get_doc(obj);
-
+  return_if_fail(conf != NULL);
   modbus_client_channel_t* channel = modbus_client_channel_create(conf->root);
   ASSERT_EQ(channel->unit_id, 1);
   ASSERT_STREQ(channel->name, "write_registers");